fix(types): Resolve Long/Int type mismatches feat(surgery): Implement Add Surgery workflow fix(surgery): Note Add/Delete Surgery bugs

+++ src/main/kotlin/com/medichart/model/Surgery.kt
@@ -6,7 +6,7 @@ import java.time.LocalDate
  * Represents a past surgery.
  */
 data class Surgery(
-    val id: Int = 0,    // Database ID
+    val id: Long = 0,    // Database ID
     val name: String,
     val date: LocalDate?,  // Uses String for date simplicity
     val surgeon: String?
